functions/loader-binance/requirements.txt
-i https://pypi.org/simple
binance-futures-connector==4.0.0; python_version >= '3.7'
boto3==1.29.6; python_version >= '3.7'
botocore==1.32.6; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
idna==3.6; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
pycryptodome==3.19.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pymongo[srv]==4.6.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
s3transfer==0.7.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==2.0.7; python_version >= '3.10'
websocket-client==1.6.4; python_version >= '3.8'
